python/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
blinker==1.9.0
click==8.3.0
fastapi==0.120.4
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
psycopg==3.2.12
psycopg-binary==3.2.12
pydantic==2.12.3
pydantic_core==2.41.4
pyserial==3.5
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
Werkzeug==3.1.3
